fix: styling and rendering issues with traces details
