+++ client/src/components/LlmPanel.tsx
@@ -32,7 +32,9 @@ export const LlmPanel = ({ spanUuid }: { spanUuid: string }) => {
         <Badge>{(span.duration_ms / 1_000_000_000).toFixed(3)}s</Badge>
       </div>
       {span.messages.length > 0 && renderMessagesContainer(span.messages)}
-      {renderData(span.data)}
+      {renderData({
+        value: span.data,
+      })}
     </div>
   );
 };
